Moved rasa to backend, general restructuring (#51)
* Moved rasa to backend, general restructuring

* Modify workflows

* Install rasa in workflow

* Added working directory to rasa install workflow
