+++ .github/workflows/main.yml
@@ -47,3 +47,19 @@ jobs:
           build: npm run build
           start: npm start
           working-directory: backend
+  rasa:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python 3.7
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.7
+      - name: Install rasa
+        run: |
+          python -m pip install -U pip
+          pip install -r requirements.txt
+        working-directory: backend/rasa
+      - name: Train rasa
+        run: rasa train
+        working-directory: backend/rasa
