+++ backend/controllers/trollbotRouter.js
@@ -1,15 +1,9 @@
 const trollbotRouter = require('express').Router()
-const {botAnswer, getMessages, clearMessages} = require('./trollbotAnswerController')
-
-
-
-
-
+const {botAnswer, getMessages, clearMessages} = require('../services/trollbot')
 
 trollbotRouter.get('/', (req, res) => {
   res.json(getMessages())
 })
-
 
 trollbotRouter.post('/', async (req, res) => {
   const response = await botAnswer(req.body)
