--- backend/controllers/userIntentController.js
@@ -1,9 +1,0 @@
-// swap this when needed. must export a getIntent function.
-const intentModule = require('../userIntentModules/userIntentFromJson')
-// ^ swap this
-
-const getIntent = (message) => {
-  return intentModule(message)
-}
-
-module.exports = getIntent
