+++ backend/package.json
@@ -7,6 +7,8 @@
     "start": "cross-env NODE_ENV=production node index.js",
 	"start-staging": "set API_URL=/trollbot/api/trollbot&& set PUBLIC_URL=/trollbot&& cross-env NODE_ENV=production node index.js",
     "dev": "cross-env NODE_ENV=development nodemon index.js",
+    "rasa": "cd rasa/ && rasa run",
+    "all": "npm-run-all --parallel dev rasa",
     "test": "cross-env NODE_ENV=test jest --verbose --runInBand --testPathPattern=/tests",
     "start:test": "cross-env NODE_ENV=test node index.js",
     "cy:open": "cypress open",
@@ -24,6 +26,7 @@
     "cors": "*",
     "dotenv": "^10.0.0",
     "express": "*",
+    "npm-run-all": "^4.1.5",
     "path": "^0.12.7",
     "querystring": "^0.2.0"
   },
