+++ backend/rasa/requirements.txt
@@ -1,0 +1,1 @@
+rasa==2.7.1
