+++ backend/tests/trollbot-answer.test.js
@@ -1,27 +1,26 @@
-const trollbot = require('../controllers/trollbotAnswerController')
-const { getResponse } = require('../controllers/trollbotAnswerController')
+const trollbot = require('../services/trollbot')
 
 afterEach(() => {
   trollbot.clearMessages()
 })
 
 test('Opening message returns correct bot response', async () => {
-  const response = await getResponse('hello')
+  const response = await trollbot.getResponse('hello')
   expect(['Hi!', 'Hey!', 'Hello!', 'Howdy!']).toContain(response[2].body)
 })
 
 test('Closing message returns correct bot response', async () => {
-  const response = await getResponse('bye')
+  const response = await trollbot.getResponse('bye')
   expect(['Goodbye!', 'Cya!', 'So long!']).toContain(response[2].body)
 })
 
 test('Question returns correct bot response', async () => {
-  const response = await getResponse('hmm?')
+  const response = await trollbot.getResponse('hmm?')
   expect(['What did you ask?', 'Why are you asking me?', 'Any other questions?']).toContain(response[2].body)
 })
 
 test('Other message returns correct bot response', async () => {
-  const response = await getResponse('hshjsksh')
+  const response = await trollbot.getResponse('hshjsksh')
   expect('Could not find information on this band.').toContain(response[2].body)
 })
