--- rasa/data/nlu.yml
@@ -1,156 +1,0 @@
-version: "2.0"
-
-nlu:
-- intent: opening
-  examples: |
-    - hey
-    - hello
-    - hi
-    - hello there
-    - good morning
-    - good evening
-    - moin
-    - hey there
-    - let's go
-    - hey dude
-    - goodmorning
-    - goodevening
-    - good afternoon
-
-- intent: closing
-  examples: |
-    - good afternoon
-    - cu
-    - good by
-    - cee you later
-    - good night
-    - bye
-    - goodbye
-    - have a nice day
-    - see you around
-    - bye bye
-    - see you later
-
-- intent: question
-  examples: |
-    - what
-    - why
-    - who are you
-    - how come
-    - when
-    - which one
-    - are you sure
-    - can you help me
-    - what is going on
-    - hmm?
-
-- intent: likes_artist
-  examples: |
-    - my favourite band is [Abba]{"entity": "artist"}
-    - i like [Queen]{"entity": "artist"}
-    - i really like [Abba]{"entity": "artist"}
-    - i enjoy listening to [Queen] {"entity": "artist"}
-    - i love [Abba]{"entity": "artist"}
-    - [Queen]{"entity": "artist"} rocks
-    - [Abba]{"entity": "artist"} is so good
-    - [Adele]{"entity": "artist"} is such a great singer
-    - [Abba]{"entity": "artist"} has a great sound
-
-- intent: dislikes_artist
-  examples: |
-    - my least favourite band is [Abba]{"entity": "artist"}
-    - i don't like [Abba]{"entity": "artist"}
-    - i really don't like [Queen]{"entity": "artist"}
-    - i hate [Queen]{"entity": "artist"}
-    - [Abba]{"entity": "artist"} sucks
-    - [Abba]{"entity": "artist"} is so bad
-    - [Queen]{"entity": "artist"} is so overrated
-    - [Abba]{"entity": "artist"} is garbage
-
-- lookup: artist
-  examples: |
-    - Abba
-    - Adele
-    - Aphex Twin
-    - Arcade Fire
-    - Ariana Grande
-    - The Beatles
-    - Black Sabbath
-    - Bob Dylan
-    - Bruce Springsteen
-    - The Clash
-    - Common
-    - Cream
-    - Daft Punk
-    - David Bowie
-    - The Doors
-    - Drake
-    - Dream Theater
-    - Elton John
-    - Elvis Presley
-    - Eminem
-    - Foo Fighters
-    - Frank Sinatra
-    - Frank Zappa
-    - George Michael
-    - Genesis
-    - Gorillaz
-    - Green Day
-    - Guns N' Roses
-    - Gwen Stefani
-    - HIM
-    - Hoobastank
-    - Iron Maiden
-    - Jackson 5
-    - Jay-Z
-    - Jimi Hendrix
-    - Joy Division
-    - Justin Bieber
-    - Kanye West
-    - Katy Perry
-    - Kendrick Lamar
-    - Kiss
-    - Lady Gaga
-    - Led Zeppelin
-    - Lil Wayne
-    - Lorde
-    - Madonna
-    - Marilyn Manson
-    - Massive Attack
-    - Metallica
-    - Michael Jackson
-    - Muse
-    - Neil Young
-    - Nicki Minaj
-    - Nirvana
-    - Opeth
-    - Outkast
-    - Phil Collins
-    - Pink Floyd
-    - Porcupine Tree
-    - Portishead
-    - Prince
-    - Public Enemy
-    - Queen
-    - Rammstein
-    - Radiohead
-    - Red Hot Chili Peppers
-    - Rihanna
-    - The Rolling Stones
-    - Rush
-    - Sex Pistols
-    - Slayer
-    - Snoop Dogg
-    - Soundgarden
-    - Tame Impala
-    - Taylor Swift
-    - Tool
-    - Tom Waits
-    - U2
-    - Uriah Heep
-    - Van Halen
-    - Wu-Tang Clan
-    - The Who
-    - Yes
-    - The Zombies
-    - 65daysofstatic
