--- rasa/data/rules.yml
@@ -1,13 +1,0 @@
-version: "2.0"
-
-rules:
-
-- rule: Say goodbye anytime the user says goodbye
-  steps:
-  - intent: closing
-  - action: utter_closing
-
-- rule: Ask the user to rephrase whenever they send a message with low NLU confidence
-  steps:
-  - intent: nlu_fallback
-  - action: utter_please_rephrase
