--- rasa/domain.yml
@@ -1,64 +1,0 @@
-version: "2.0"
-
-intents:
-  - opening
-  - closing
-  - affirm
-  - deny
-  - mood_great
-  - mood_unhappy
-  - bot_challenge
-  - question
-  - likes_artist
-  - dislikes_artist
-
-entities:
-  - artist
-
-slots: 
-  artist:
-    type: text
-    auto_fill: true
-
-# forms:
-#   artist_form:
-#     required_slots:
-#       artist:
-#         - type: from_entity
-#           entity: artist
-          
-responses:
-  utter_opening:
-  - text: "Hey!"
-  - text: "Hi!"
-  - text: "Hello!"
-  - text: "Howdy!"
-
-  utter_closing:
-  - text: "Bye!"
-  - text: "Cya!"
-  - text: "So long!"
-
-  utter_answer:
-  - text: "Any other questions?"
-  - text: "What did you ask?"
-  - text: "Why are you asking me?"
-
-  utter_artist_dismissal:
-  - text: "Ew, {artist}, really? I guess there's no accounting for taste."
-  - text: "{artist}? Not doing it for me."
-  - text: "{artist}? I'd rather listen to nails on a chalkboard."
-
-  utter_artist_praise:
-  - text: "What? {artist} is wonderful!"
-  - text: "{artist} rocks and your opinion is dismissed."
-  - text: "Really? IMO, {artist} is one of the greats."
-
-  utter_please_rephrase:
-  - text: "I don't understand."
-  - text: "Can you repeat that with different words?"
-  - text: "I'm sorry, I don't follow. Can you be more specific?"
-
-session_config:
-  session_expiration_time: 60
-  carry_over_slots_to_new_session: true
